requirements.txt
pandas
google-play-scraper
app-store-scraper
groq
python-dotenv
streamlit
plotly
